Move triangulation code to its own module
