+++ computational_geometry/src/lib.rs
@@ -11,5 +11,6 @@ pub mod line_segment;
 pub mod point;
 pub mod polygon;
 pub mod triangle;
+pub mod triangulation;
 pub mod vertex;
 pub mod vertex_map;
